Fixed issue with setps (oby, gby) and dfe dependency
When placing a dfe record all the setps (oby,gby) between the use
of the value and its computation, the dfe is added as a gby/oby
dependent. When placing predicates that are imported into a dt and
these contain subqs e.g. col = scalar subq where the scalar subq
is independent of the dt being placed, then disable subq caching.
If kept caching on but placed the subq before the dt the subq would
be generated twice because renamed once when importing into the dt