Fixed isue with dfe in loc results
The dfes in loc results  required can be copies of the canonical
dfe for the tree that is by sqlo df elt.  So in placing, to detect
where a dfe in fact is, go to the copy from the canonical.  When
generating code, go from the copy to the canonical so that if a ssl
is placed for the remote, the ssl is found when getting the dfe by
the exp tree