Fixed issue with outerjoin
Can be in oj a ij (b oj c) on qq that qq must be in top where and
not in the c oj cond. For this use not res but the ot enclosing
where cond field. Used for nested cases. A tree with ojs can become
a dt and that dt gets the top where conds, not the top dt.