Fixed issue with heterogeneous data column leading to range assert
Usually this code block deals with data column with only numbers, dates
or any, but if the data col has reference to GROUP BY & DISTINCT it
eventually referenced RDF objects and it becomes heterogeneous.

This case was not handled, leading to a range assert where in fact the
column structure is fine.