Fixed issue with restriction on number of deleted triples (fixes #1164)
The call that was supposed to signal an error was called in the wrong
place.

When using a delete statemtent like this:

    DELETE WHERE { graph <mygraph-uri> { ?p ?q ?r } }

Virtuoso internally uses a construct to generate a dict of the triples
to be deleted.

Virtuoso uses settings like ResultSetMaxRows and MaxConstructTriples to
limit the amount of triples can be used by a single SPARQL call.

You can configure the maximum number of triples in a construct by editing
the virtuoso.ini:

      [SPARQL]
      MaxConstructTriples = 10000  ; if not set, use ResultSetMaxRows as fallback