Added new MaxTempDBPages setting to virtuoso.ini
This setting limits the maximum size of the TempDatabase file on disk to a
fixed amount of used pages. If the engine detects this limit has been
exceeded, the query that was filling up the TempDatabase will return an
error and its transaction is killed.

  [TempDatabase]
  MaxTempDBPages = 50G  ; Limit size of the TempDatabase to 50GB on disk

Values ending in B, K, M, G will be converted to number of pages that
can fit into that size. Plain values are interpreted as number of pages.

For example 50G = (50 * 1024 * 1024 * 1024) / 8192 = 6553600 pages

The default value is 0 which means that the TempDatabase file can fill
up the whole disk partition it is located on.