Fixed maximum lenght of an long url to 64K
Some RDF datasets have extreme long URIs that exceed the Virtuoso internal
maximum size of 1900 bytes and thus need to be shortened to avoid errors
such as:

    SR133: Can not set NULL to not nullable column 'DB.DBA.RDF_QUAD.S'

The ShortenLongURIs parameter is a Virtuoso configuration setting which
shorten extremely long URIs in datasets when loading with the RDF Bulk
Loader. When enabled, Virtuoso will truncate the long url and appends
a checksum of the original url to ensure it is still unique.

Add the following setting to your virtuoso.ini file and restart your
Virtuoso instance to enable this feature:

    [SPARQL]
    ShortenLongURIs = 1